requirements.txt
# Зависимости для Telegram Bot Manager
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
asyncio
sqlite3
